refact : 투표 조회 메서드 도메인 객체로 마이그레이션
