+++ src/main/java/co/kr/jurumarble/vote/repository/VoteEntityRepositoryImpl.java
@@ -1,6 +1,5 @@
 package co.kr.jurumarble.vote.repository;
 
-import co.kr.jurumarble.exception.vote.IllegalVoteTypeException;
 import co.kr.jurumarble.user.enums.ChoiceType;
 import co.kr.jurumarble.user.enums.GenderType;
 import co.kr.jurumarble.user.enums.MbtiType;
@@ -24,9 +23,7 @@ import javax.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static co.kr.jurumarble.user.domain.QUser.user;
 import static co.kr.jurumarble.vote.domain.QVote.vote;
@@ -76,29 +73,6 @@ public class VoteEntityRepositoryImpl implements VoteEntityRepository {
                 .fetch();
     }
 
-    // 1. 원하는 조건을 걸어서 투표 컨텐츠 없이 투표를 찾는다.
-    // 2. 찾은 투표의 아이디와 타입으로 투표 컨텐츠를 찾는다.
-    // 3. 찾은 투표 컨텐츠와 투표 데이터를 이어서 완전하게 만든다.
-    public Slice<VoteData> findVoteDataWithPopularity(String keyword, Pageable pageable) {
-
-        List<VoteCommonData> voteCommonDataList = getVoteCommonDataByPopular(keyword, pageable);
-
-        List<Long> normalVoteIds = getNormalVoteIdsFromFindVoteCommonDataList(voteCommonDataList);
-        List<VoteContent> voteContents = findVoteContentsByNormalVoteIds(normalVoteIds);
-
-        List<Long> drinkVoteIds = getDrinkVoteIdsFromFindVoteCommonDataList(voteCommonDataList);
-        List<VoteDrinkContent> voteDrinkContents = findVoteContentsByDrinkVoteIds(drinkVoteIds);
-
-        Map<Long, VoteContent> voteContentsMap = voteContents.stream()  // List를 순회하면 성능이 안나오므로 <voteId, VoteConent> 로 이루어진 Map을 만듬
-                .collect(Collectors.toMap(VoteContent::getVoteId, voteContent -> voteContent));// ex) <1, {voteContent}>
-
-        Map<Long, VoteDrinkContent> voteDrinkContentMap = voteDrinkContents.stream()
-                .collect(Collectors.toMap(VoteDrinkContent::getVoteId, voteDrinkContent -> voteDrinkContent));
-
-        List<VoteData> voteData = getFindVoteDataList(voteCommonDataList, voteContentsMap, voteDrinkContentMap);
-
-        return getSlice(voteData, pageable.getPageSize(), pageable.getPageSize(), pageable);
-    }
 
     private SliceImpl<VoteData> getSlice(List<VoteData> voteData, int pageable, int pageable1, Pageable pageable2) {
         boolean hasNext = false;
@@ -110,37 +84,6 @@ public class VoteEntityRepositoryImpl implements VoteEntityRepository {
 
         return new SliceImpl<>(voteData, pageable2, hasNext);
     }
-
-//    private List<VoteCommonData> getVoteCommonDataByPopular(String keyword, Pageable pageable) {
-//        int pageNo = pageable.getPageNumber();
-//        int pageSize = pageable.getPageSize();
-//
-//        BooleanExpression keywordExpression = getKeywordExpression(keyword);
-//
-//        return jpaQueryFactory
-//                .select(
-//                        Projections.bean(
-//                                VoteCommonData.class,
-//                                vote.id.as("voteId"),
-//                                vote.postedUserId,
-//                                vote.title,
-//                                vote.detail,
-//                                vote.filteredGender,
-//                                vote.filteredAge,
-//                                vote.filteredMbti,
-//                                voteResult.id.count().as("votedCount"),
-//                                vote.voteType)
-//                )
-//                .from(vote)
-//                .leftJoin(voteResult).on(vote.id.eq(voteResult.voteId))
-//                .where(keywordExpression)
-//                .groupBy(vote.id)
-//                .orderBy(voteResult.id.count().desc())
-//                .offset(pageNo * pageSize)
-//                .limit(pageSize)
-//                .fetch();
-//    }
-
 
 
 
@@ -160,53 +103,6 @@ public class VoteEntityRepositoryImpl implements VoteEntityRepository {
                 .fetch();
     }
 
-    private List<VoteData> getFindVoteDataList(List<VoteCommonData> voteCommonDataList, Map<Long, VoteContent> voteContentsMap, Map<Long, VoteDrinkContent> voteDrinkContentMap) {
-        return voteCommonDataList.stream()
-                .map(voteCommonData -> {
-                    if (voteCommonData.getVoteType().equals(VoteType.NORMAL)) {
-                        VoteContent voteContent = voteContentsMap.get(voteCommonData.getVoteId());
-                        return generateNormalVoteData(voteCommonData, voteContent);
-                    }
-                    if (voteCommonData.getVoteType().equals(VoteType.DRINK)) {
-                        VoteDrinkContent voteDrinkContent = voteDrinkContentMap.get(voteCommonData.getVoteId());
-                        return generateDrinkVoteData(voteCommonData, voteDrinkContent);
-                    }
-                    throw new IllegalVoteTypeException();
-                }).collect(Collectors.toList());
-    }
-
-    private static VoteData generateDrinkVoteData(VoteCommonData voteCommonData, VoteDrinkContent voteDrinkContent) {
-        return VoteData.builder()
-                .voteId(voteCommonData.getVoteId())
-                .postedUserId(voteCommonData.getPostedUserId())
-                .title(voteCommonData.getTitle())
-                .detail(voteCommonData.getDetail())
-                .filteredGender(voteCommonData.getFilteredGender())
-                .filteredAge(voteCommonData.getFilteredAge())
-                .imageA(voteDrinkContent.getDrinkAImage())
-                .imageB(voteDrinkContent.getDrinkBImage())
-                .titleA(voteDrinkContent.getDrinkAName())
-                .titleB(voteDrinkContent.getDrinkBName())
-                .votedCount(voteCommonData.getVotedCount())
-                .region(voteDrinkContent.getRegion())
-                .build();
-    }
-
-    private VoteData generateNormalVoteData(VoteCommonData voteCommonData, VoteContent voteContent) {
-        return VoteData.builder()
-                .voteId(voteCommonData.getVoteId())
-                .postedUserId(voteCommonData.getPostedUserId())
-                .title(voteCommonData.getTitle())
-                .detail(voteCommonData.getDetail())
-                .filteredGender(voteCommonData.getFilteredGender())
-                .filteredAge(voteCommonData.getFilteredAge())
-                .imageA(voteContent.getImageA())
-                .imageB(voteContent.getImageB())
-                .titleA(voteContent.getTitleA())
-                .titleB(voteContent.getTitleB())
-                .votedCount(voteCommonData.getVotedCount())
-                .build();
-    }
 
     @Override
     public Optional<VoteData> findVoteDataByVoteId(Long voteId) {
